feat: format and convert Sats to BTC

+++ domain/src/main/java/com/muniz/lightningclient/domain/model/Node.kt
@@ -4,7 +4,7 @@ data class Node(
     val publicKey: String,
     val alias: String,
     val channels: Int,
-    val capacity: Long,
+    val capacity: Double,
     val firstSeen: Long,
     val updatedAt: Long,
     val city: City?,
